requirements.txt
grpcio
grpcio-tools
protobuf
fastapi
uvicorn
numpy
pandas
pytest
